wip shell completions
