+++ .gitignore
@@ -1,4 +1,5 @@
 /target
+/completions
 result*
 docs/book
 .nixos-test-history
