+++ Cargo.toml
@@ -3,10 +3,16 @@ name = "cargo-reaper"
 version = "0.2.0"
 authors = ["Cloud Scythe Labs <cloudscythelabs@gmail.com>"]
 description = "A Cargo plugin for developing REAPER extension plugins with Rust."
+build = "build.rs"
 edition = "2024"
 license = "MIT"
 readme = "README.md"
 publish = true
+
+[build-dependencies]
+clap = { version = "4.5", features = ["derive", "env"] }
+clap_complete = "4.5"
+humantime = "2.2"
 
 [dependencies]
 anyhow = "1"
